refactor: package.json 내용 수정 및 api 폴더명 변경
