--- src/services/index.js
@@ -1,1 +1,0 @@
-export { default as API } from './api';
