--- src/utils/index.js
@@ -1,1 +1,0 @@
-export { default as request } from './request';
